encode the document
