+++ rust/signed_doc/src/builder.rs
@@ -2,32 +2,29 @@
 
 /// An implementation of [`CborMap`].
 mod cbor_map;
-/// COSE protected header as per [RFC 8152](https://datatracker.ietf.org/doc/html/rfc8152#autoid-8),
-/// but with some fields omitted when unused by Catalyst and some fields specialized for
-/// it.
-mod cose_header;
-/// Signed payload with signatures as described in
-/// [section 2 of RFC 8152](https://datatracker.ietf.org/doc/html/rfc8152#section-2).
-/// Specialized for Catalyst Signed Document (e.g. no support for unprotected headers).
-mod cose_sign;
+/// COSE format utils.
+mod cose;
 
 use std::convert::Infallible;
 
 use catalyst_types::catalyst_id::CatalystId;
 use cbor_map::CborMap;
-use cose_header::{make_metadata_header, make_signature_header};
-use cose_sign::make_tbs_data;
-
-use crate::CatalystSignedDocument;
+use cose::{
+    encode_cose_sign, make_cose_signature, make_metadata_header, make_signature_header,
+    make_tbs_data,
+};
 
 pub type EncodeError = minicbor::encode::Error<Infallible>;
 
 /// Catalyst Signed Document Builder.
 #[derive(Debug)]
 pub struct Builder {
+    /// Mapping from encoded keys to encoded values.
     metadata: CborMap,
+    /// Encoded document content.
     content: Vec<u8>,
-    signatures: Vec<(CatalystId, Vec<u8>)>,
+    /// Encoded COSE Signatures.
+    signatures: Vec<Vec<u8>>,
 }
 
 impl Builder {
@@ -66,25 +63,28 @@ impl Builder {
     pub fn add_signature(
         mut self, sign_fn: impl FnOnce(Vec<u8>) -> Vec<u8>, kid: CatalystId,
     ) -> Result<Self, EncodeError> {
+        // Question: maybe this should be cached (e.g. frozen once filled)?
         let metadata_header = make_metadata_header(&self.metadata);
 
         let kid_str = kid.to_string().into_bytes();
         let signature_header = make_signature_header(kid_str.as_slice())?;
 
-        // Question: maybe this should be cached?
         let tbs_data = make_tbs_data(&metadata_header, &signature_header, &self.content)?;
+        let signature_bytes = sign_fn(tbs_data);
 
-        let signature = sign_fn(tbs_data);
-
-        self.signatures.push((kid, signature));
+        let signature = make_cose_signature(&signature_header, &signature_bytes)?;
+        self.signatures.push(signature);
 
         Ok(self)
     }
 
-    /// Build a signed document with the collected error report.
+    /// Build a CBOR-encoded signed document with the collected error report.
     /// Could provide an invalid document.
     #[must_use]
-    pub fn build(self) -> CatalystSignedDocument {
-        self.0.into()
+    pub fn build<W: minicbor::encode::Write>(
+        &self, e: &mut minicbor::Encoder<W>,
+    ) -> Result<(), minicbor::encode::Error<W::Error>> {
+        let metadata_header = make_metadata_header(&self.metadata);
+        encode_cose_sign(e, &metadata_header, &self.content, &self.signatures)
     }
 }
